dbt_dag/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Cosmos for Airflow-dbt integration
astronomer-cosmos[celery]>=1.4.0

# dbt packages for Snowflake
# Let dbt-snowflake automatically pull in compatible dbt-core version
dbt-snowflake>=1.10.0

# Snowflake provider for Airflow
apache-airflow-providers-snowflake